+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Event Query Language - Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+# Version 0.9.19
+
+ _Released 2023-10-10_
+
+### Added
+
+* Added IPv6 support for CidrMatch
+* Removed the regex support for testing CidrMatch in favor of the native ipaddress module testing
+
 # Version 0.9.18
 
  _Released 2023-09-01_